adjust stats api
